Struct ryml::NodeRef

pub struct NodeRef<'a, 't, 'k, T>
where
    T: AsRef<Tree<'a>> + 't,
    'a: 't,
{ /* private fields */ }

A reference to a node in the tree.

Implementations

impl<'a, 't, T> NodeRef<'a, 't, '_, T>
where T: AsRef<Tree<'a>> + 't, 'a: 't,

pub fn tree<'r>(&'r self) -> &'t Tree<'a>

Get the tree the node belongs to.

pub fn data<'r>(&'r self) -> Option<&NodeData<'t>>

Get the node data, if it exists and is still valid.

Note: If the tree has been reordered, this may not return the same node as it orginally pointed to. However, it is guaranteed to still point to a valid node on the tree.

pub unsafe fn data_unchecked<'r>(&'r self) -> &'t NodeData<'t>

Get the node data, assuming it still exists and is valid.

Safety

Calling this method if the node no longer exists is undefined behaviour.

pub fn is_valid(&self) -> bool

Check if the node reference points to a valid node.

pub fn is_seed(&self) -> bool

Check if the node reference holds a seed for a non-existent node.

pub fn node_type(&self) -> Result<NodeType, Error>

Get the node type, if it exists.

pub fn node_type_as_str(&self) -> Result<&str, Error>

Get the node type name, if it exists.

pub fn key(&self) -> Result<&str, Error>

Get the node key, if it exists.

pub fn key_tag(&self) -> Result<&str, Error>

Get the tag on the node key, if it exists.

pub fn key_ref(&self) -> Result<&str, Error>

Get the reference on the node key, if it exists.

pub fn key_anchor(&self) -> Result<&str, Error>

Get the anchor on the node key, if it exists.

pub fn key_scalar(&self) -> Result<&NodeScalar<'_>, Error>

Get the scalar data of the node key, if it exists.

pub fn val(&self) -> Result<&str, Error>

Get the node value, if it exists.

pub fn val_tag(&self) -> Result<&str, Error>

Get the tag on the node value, if it exists.

pub fn val_ref(&self) -> Result<&str, Error>

Get the reference on the node value, if it exists.

pub fn val_anchor(&self) -> Result<&str, Error>

Get the anchor on the node value, if it exists.

pub fn val_scalar(&self) -> Result<&NodeScalar<'_>, Error>

Get the scalar data of the node value, if it exists.

pub fn is_stream(&self) -> Result<bool, Error>

Check if the node is a stream

pub fn is_doc(&self) -> Result<bool, Error>

Check if the node is a doc

pub fn is_container(&self) -> Result<bool, Error>

Check if the node is a container

pub fn is_map(&self) -> Result<bool, Error>

Check if the node is a map

pub fn is_seq(&self) -> Result<bool, Error>

Check if the node is a seq

pub fn has_val(&self) -> Result<bool, Error>

Check if the node has a value

pub fn has_key(&self) -> Result<bool, Error>

Check if the node has a key

pub fn is_val(&self) -> Result<bool, Error>

Check if the node is a value

pub fn is_keyval(&self) -> Result<bool, Error>

Check if the node is a keyval

pub fn has_key_tag(&self) -> Result<bool, Error>

Check if the node has a key tag

pub fn has_val_tag(&self) -> Result<bool, Error>

Check if the node has a value tag

pub fn has_key_anchor(&self) -> Result<bool, Error>

Check if the node has a key anchor

pub fn has_val_anchor(&self) -> Result<bool, Error>

Check if the node has a value anchor

pub fn has_anchor(&self, anchor: &str) -> Result<bool, Error>

Check if the node has a anchor

pub fn is_anchor(&self) -> Result<bool, Error>

Check if the node is a anchor

pub fn is_key_ref(&self) -> Result<bool, Error>

Check if the node is a key ref

pub fn is_val_ref(&self) -> Result<bool, Error>

Check if the node is a value ref

pub fn is_ref(&self) -> Result<bool, Error>

Check if the node is a ref

pub fn is_anchor_or_ref(&self) -> Result<bool, Error>

Check if the node is a anchor or ref

pub fn is_key_quoted(&self) -> Result<bool, Error>

Check if the node is a key quoted

pub fn is_val_quoted(&self) -> Result<bool, Error>

Check if the node is a value quoted

pub fn is_quoted(&self) -> Result<bool, Error>

Check if the node is a quoted

pub fn parent_is_seq(&self) -> Result<bool, Error>

Check if the parent is a sequence

pub fn parent_is_map(&self) -> Result<bool, Error>

Check if the parent is a map

pub fn is_empty(&self) -> Result<bool, Error>

Returns true if the name and value are empty and the node has no children.

pub fn is_root(&self) -> Result<bool, Error>

Returns true if the node is the root of the tree.

pub fn has_parent(&self) -> Result<bool, Error>

Returns true if the node has a parent.

pub fn has_child(&self, key: &str) -> Result<bool, Error>

Returns true if the node is a map and has a child with the given key.

pub fn has_child_at(&self, pos: usize) -> Result<bool, Error>

Returns true if the node is a sequence and has a child at the given position.

pub fn has_children(&self) -> Result<bool, Error>

Returns true if the node has children.

pub fn has_sibling(&self, key: &str) -> Result<bool, Error>

Returns true if the node has a sibling with the given key.

pub fn has_sibling_at(&self, index: usize) -> Result<bool, Error>

Returns true if the node has a sibling at the given position.

pub fn has_siblings(&self) -> Result<bool, Error>

Returns true if the node has siblings.

Note: This corresponds to has_other_siblings() in the C++ API, as the plain has_siblings() function always returns true by counting the node itself, which seems rather pointless and is not the obvious meaning of a method by the name.

pub fn parent<'r>(&'r self) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Returns a NodeRef to the parent node, if it exists.

pub fn prev_sibling<'r>( &'r self, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Returns a NodeRef to the previous sibling, if it exists.

pub fn next_sibling<'r>( &'r self, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Returns a NodeRef to the next sibling, if it exists.

pub fn num_children(&self) -> Result<usize, Error>

Get the number of children of the node.

pub fn child_pos(&self, child: &NodeRef<'a, 't, '_, T>) -> Result<usize, Error>

Get the position of the child node in this node’s children.

pub fn first_child<'r>( &'r self, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the first child of this node, if it exists.

pub fn last_child<'r>( &'r self, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the last child of this node, if it exists.

pub fn child_at<'r>( &'r self, pos: usize, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the child of this node at the given position, if it exists.

pub fn find_child<'r>( &'r self, key: &str, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the child of this node with the given key, if it exists.

pub fn first_sibling<'r>( &'r self, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the first sibling of this node, if it exists.

pub fn last_sibling<'r>( &'r self, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the last sibling of this node, if it exists.

pub fn sibling_at<'r>( &'r self, pos: usize, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the sibling of this node at the given position, if it exists.

pub fn find_sibling<'r>( &'r self, key: &str, ) -> Result<NodeRef<'a, 't, '_, &'t Tree<'a>>, Error>

Get a NodeRef to the sibling of this node with the given key, if it exists.

pub fn get<'r, 'k2, S: Into<Seed<'k2>>>( &'r self, lookup: S, ) -> Result<NodeRef<'a, 't, 'k2, &'t Tree<'a>>, Error>

Get a NodeRef to a child of this node by its given key (if this node is a map) or given position (if this node is a sequence).

Unlike get_mut, this method will return a NodeNotFound error if the child node does not exist. It will also return this error if the current node does not exist.

pub fn iter(&self) -> Result<NodeIterator<'a, 't, '_, &'t Tree<'a>>, Error>

Iterate over the children of this node, if it exists and is valid.

impl<'a, 't> NodeRef<'a, 't, '_, &'t mut Tree<'a>>

pub fn tree_mut<'r>(&'r mut self) -> &'t mut Tree<'a>

Get a mutable reference to the tree the node belongs to.

pub fn data_mut<'r>(&'r mut self) -> Option<&'t mut NodeData<'t>>

Get a mutable reference to the node data, if it exists and is still valid.

pub unsafe fn data_unchecked_mut<'r>(&'r mut self) -> &'t mut NodeData<'t>

Get a mutable reference to the node data, assuming it still exists and is valid.

Safety

Calling this method if the node no longer exists is undefined behaviour and should be used with the utmost caution.

pub fn parent_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Returns a mutable NodeRef to the parent node, if it exists.

pub fn prev_sibling_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Returns a mutable NodeRef to the previous sibling, if it exists.

pub fn next_sibling_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Returns a mutable NodeRef to the next sibling, if it exists.

pub fn first_child_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the first child of this node, if it exists.

pub fn last_child_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the last child of this node, if it exists.

pub fn child_at_mut<'r>( &'r mut self, pos: usize, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the child of this node at the given position, if it exists.

pub fn find_child_mut<'r>( &'r mut self, key: &str, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the child of this node with the given key, if it exists.

pub fn first_sibling_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the first sibling of this node, if it exists.

pub fn last_sibling_mut<'r>( &'r mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the last sibling of this node, if it exists.

pub fn sibling_at_mut<'r>( &'r mut self, pos: usize, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the sibling of this node at the given position, if it exists.

pub fn find_sibling_mut<'r>( &'r mut self, key: &str, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to the sibling of this node with the given key, if it exists.

pub fn change_type(&mut self, node_type: NodeType) -> Result<bool, Error>

Change the type of the node, resetting its contents if necessary and returning whether the change was possible.

pub fn set_type_flags(&mut self, more_flags: NodeType) -> Result<(), Error>

Set flags on the node.

pub fn set_key(&mut self, key: &str) -> Result<(), Error>

Sets the node’s key.

pub fn set_val(&mut self, value: &str) -> Result<(), Error>

Sets the node’s value.

pub fn set_key_tag(&mut self, v: &str) -> Result<(), Error>

Set the tag on the node key.

pub fn set_val_tag(&mut self, v: &str) -> Result<(), Error>

Set the tag on the node val.

pub fn set_key_anchor(&mut self, v: &str) -> Result<(), Error>

Set the anchor on the node key.

pub fn set_val_anchor(&mut self, v: &str) -> Result<(), Error>

Set the anchor on the node val.

pub fn set_key_ref(&mut self, v: &str) -> Result<(), Error>

Set the ref on the node key.

pub fn set_val_ref(&mut self, v: &str) -> Result<(), Error>

Set the ref on the node val.

pub fn clear(&mut self) -> Result<(), Error>

Empties the node and removes any children.

pub fn clear_key(&mut self) -> Result<(), Error>

Clears the node key, if it exists.

pub fn clear_val(&mut self) -> Result<(), Error>

Clears the node value, if it exists.

pub fn clear_children(&mut self) -> Result<(), Error>

Clear the node’s children, if it exists and has any.

pub fn insert_child<R: AsRef<Tree<'a>>>( &mut self, after: NodeRef<'a, 't, '_, R>, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Insert a new node as a child of this node, returning a NodeRef to the new node.

pub fn prepend_child( &mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Insert a new node as the first child of this node, returning a NodeRef to the new node.

pub fn append_child( &mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Insert a new node as the last child of this node, returning a NodeRef to the new node.

pub fn insert_sibling<R: AsRef<Tree<'a>>>( &mut self, after: NodeRef<'a, 't, '_, R>, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Insert a new node as a sibling of this node, returning a NodeRef to the new node.

pub fn prepend_sibling( &mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Insert a new node as the first sibling of this node, returning a NodeRef to the new node.

pub fn append_sibling( &mut self, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Insert a new node as the last sibling of this node, returning a NodeRef to the new node.

pub fn remove_child( &mut self, child: NodeRef<'a, 't, '_, &'t mut Tree<'a>>, ) -> Result<(), Error>

Remove the given child from this node.

pub fn remove_child_at(&mut self, pos: usize) -> Result<(), Error>

Remove the child at the given index from this node.

pub fn remove_child_with_key(&mut self, key: &str) -> Result<(), Error>

Remove the child with the given key from this node.

pub fn move_<R: AsRef<Tree<'a>>>( &mut self, after: NodeRef<'a, 't, '_, R>, ) -> Result<(), Error>

Change the node’s position within its parent.

pub fn move_to_parent<R: AsRef<Tree<'a>>>( &mut self, parent: NodeRef<'a, 't, '_, &'t mut Tree<'a>>, after: NodeRef<'a, 't, '_, R>, ) -> Result<(), Error>

Move the node to a different parent, which may be in a new tree.

pub fn duplicate<R: AsRef<Tree<'a>>>( &mut self, parent: NodeRef<'a, 't, '_, &'t mut Tree<'a>>, after: NodeRef<'a, 't, '_, R>, ) -> Result<NodeRef<'a, 't, '_, &'t mut Tree<'a>>, Error>

Duplicate the node under a new parent, returning a NodeRef to the new node.

pub fn duplicate_children<R: AsRef<Tree<'a>>>( &mut self, parent: NodeRef<'a, 't, '_, &'t mut Tree<'a>>, after: NodeRef<'a, 't, '_, R>, ) -> Result<(), Error>

Duplicate the node’s children under a new parent.

pub fn get_mut<'r, 'k2, S: Into<Seed<'k2>>>( &'r mut self, lookup: S, ) -> Result<NodeRef<'a, 't, 'k2, &'t mut Tree<'a>>, Error>

Get a mutable NodeRef to a child of this node by its given key (if this node is a map) or given position (if this node is a sequence).

Unlike get, this method will succeed if the node does not exist and will retain the seed for lazy assignment. It will still return a NodeNotFound error if the current node does not exist.

pub fn iter_mut( &mut self, ) -> Result<NodeIterator<'a, 't, '_, &'t mut Tree<'a>>, Error>

Iterate mutably over the children of this node, if it exists and is valid.